requirements.txt
git+https://github.com/huggingface/diffusers
torch>=2.0.0 --index-url https://download.pytorch.org/whl/cu118
torchvision --index-url https://download.pytorch.org/whl/cu118
torchaudio --index-url https://download.pytorch.org/whl/cu118
transformers
accelerate
safetensors
Flask
Pillow
psutil
